Add macOS app, UniFFI bindings, and build tooling
Add a macOS Swift frontend (app/mudd) with window/split view controllers and Xcode project config, plus a Makefile, git hook scripts, and a build script to streamline development. Introduce a small uniffi-bindgen helper crate and wiring to generate Swift bindings, update workspace members, and add build targets for bindings/app. Improve FFI surface: add a MuddError type, propagate anyhow errors through ffi exports, and add thiserror/anyhow deps. Apply multiple Rust code cleanups and small bug/format fixes across core (dicom, imaging, inference, video), and add developer documentation (AGENTS.md).

+++ core/src/imaging/roi.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use image::{GrayImage, Luma};
-use imageproc::contours::{find_contours, BorderType};
+use imageproc::contours::{BorderType, find_contours};
 
 use super::types::{ColorSpace, Frame, Roi};
 
@@ -62,8 +62,8 @@ pub fn detect_roi(frame: &Frame) -> Result<Roi> {
         let (mut max_x, mut max_y) = (0u32, 0u32);
 
         for point in &contour.points {
-            let px = point.x as u32;
-            let py = point.y as u32;
+            let px = point.x;
+            let py = point.y;
             min_x = min_x.min(px);
             min_y = min_y.min(py);
             max_x = max_x.max(px);
